+++ packages/language-service/index.js
@@ -1,8 +1,0 @@
-/**
- * @typedef {import('./lib/commands.js').SyntaxToggleParams} SyntaxToggleParams
- * @typedef {import('./lib/service-plugin.js').Commands} Commands
- */
-
-export {createMdxLanguagePlugin} from './lib/language-plugin.js'
-export {createMdxServicePlugin} from './lib/service-plugin.js'
-export {resolveRemarkPlugins} from './lib/tsconfig.js'