+++ packages/language-service/lib/index.js
@@ -1,0 +1,3 @@
+export {createMdxLanguagePlugin} from './language-plugin.js'
+export {createMdxServicePlugin} from './service-plugin.js'
+export {resolveRemarkPlugins} from './tsconfig.js'